requirements.txt
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.1
certifi==2020.4.5.1
decorator==4.4.1
Django==2.2
django-celery-results==1.2.1
django-crispy-forms==1.8.1
django-extensions==2.2.9
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.10.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
factory-boy==2.12.0
Faker==4.0.1
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
kombu==4.6.8
Markdown==3.1.1
more-itertools==8.2.0
packaging==20.3
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
